[ETCM-271] Minor refactors including Evm and Hex Utils + run services with sbt commands
