--- conseil-common/src/main/scala/tech/cryptonomic/conseil/common/ethereum/domain/Token.scala
@@ -1,14 +1,0 @@
-package tech.cryptonomic.conseil.common.ethereum.domain
-
-/**
-  * Ethereum token.
-  */
-case class Token(
-    address: String,
-    blockHash: String,
-    blockNumber: String,
-    name: String,
-    symbol: String,
-    decimals: String,
-    totalSupply: String
-)
