+++ conseil-common/src/main/scala/tech/cryptonomic/conseil/common/ethereum/rpc/EthereumRpcCommands.scala
@@ -2,7 +2,7 @@ package tech.cryptonomic.conseil.common.ethereum.rpc
 
 import io.circe.{Decoder, Encoder, HCursor, Json}
 
-import tech.cryptonomic.conseil.common.ethereum.domain.Bytecode
+import tech.cryptonomic.conseil.common.EvmDomain.Bytecode
 import tech.cryptonomic.conseil.common.rpc.RpcClient.RpcRequest
 
 /**
