+++ conseil-common/src/main/scala/tech/cryptonomic/conseil/common/mantis/MantisPersistence.scala
@@ -10,10 +10,12 @@ import slick.jdbc.PostgresProfile.api._
 
 import tech.cryptonomic.conseil.common.util.Conversion
 import tech.cryptonomic.conseil.common.util.Conversion.Syntax._
+import tech.cryptonomic.conseil.common.util.HexUtil.hexStringToBigDecimal
 import tech.cryptonomic.conseil.common.mantis.rpc.json.{Block, Log, Transaction}
 import tech.cryptonomic.conseil.common.mantis.MantisPersistence._
 import tech.cryptonomic.conseil.common.mantis.rpc.json.TransactionReceipt
-import tech.cryptonomic.conseil.common.mantis.domain.{Contract, Token}
+import tech.cryptonomic.conseil.common.EvmDomain.{Contract, Token}
+
 
 /**
   * Mantis persistence into the database using Slick.
@@ -153,7 +155,7 @@ object MantisPersistence {
           nonce = from.nonce,
           to = from.to,
           transactionIndex = from.transactionIndex,
-          value = Utils.hexStringToBigDecimal(from.value),
+          value = hexStringToBigDecimal(from.value),
           v = from.v,
           r = from.r,
           s = from.s
@@ -234,7 +236,7 @@ object MantisPersistence {
           transactionHash = from.transactionHash,
           fromAddress = from.topics(1),
           toAddress = from.topics(2),
-          value = Utils.hexStringToBigDecimal(from.data)
+          value = hexStringToBigDecimal(from.data)
         )
     }
