--- conseil-common/src/main/scala/tech/cryptonomic/conseil/common/mantis/domain/Contract.scala
@@ -1,13 +1,0 @@
-package tech.cryptonomic.conseil.common.mantis.domain
-
-/**
-  * Mantis contract.
-  */
-case class Contract(
-    address: String,
-    blockHash: String,
-    blockNumber: String,
-    isErc20: Boolean = false,
-    isErc721: Boolean = false,
-    bytecode: Bytecode
-)
