+++ conseil-common/src/main/scala/tech/cryptonomic/conseil/common/mantis/rpc/MantisRpcClient.scala
@@ -7,11 +7,12 @@ import io.circe.generic.auto._
 import org.http4s.circe.CirceEntityDecoder._
 import org.http4s.circe.CirceEntityEncoder._
 
-import tech.cryptonomic.conseil.common.mantis.domain.{Bytecode, Contract, Token}
+import tech.cryptonomic.conseil.common.EvmDomain.{Bytecode, Contract, Token}
 import tech.cryptonomic.conseil.common.rpc.RpcClient
 import tech.cryptonomic.conseil.common.mantis.rpc.MantisRpcCommands._
 import tech.cryptonomic.conseil.common.mantis.rpc.json.{Block, Transaction, TransactionReceipt}
-import tech.cryptonomic.conseil.common.mantis.Utils
+import tech.cryptonomic.conseil.common.util.HexUtil.hexToString
+import tech.cryptonomic.conseil.common.util.CryptoUtil.keccak
 
 /**
   * Mantis JSON-RPC client according to the specification at https://eth.wiki/json-rpc/API
@@ -107,7 +108,7 @@ class MantisClient[F[_]: Concurrent](
           .map(
             contract =>
               Seq("name", "symbol", "decimals", "totalSupply")
-                .map(f => EthCall.request(contract.blockNumber, contract.address, s"0x${Utils.keccak(s"$f()")}"))
+                .map(f => EthCall.request(contract.blockNumber, contract.address, s"0x${keccak(s"$f()")}"))
           )
           .flatMap(Stream.emits)
           .through(client.stream[EthCall.Params, String](batchSize = 1))
@@ -119,8 +120,8 @@ class MantisClient[F[_]: Concurrent](
                 address = contract.address,
                 blockHash = contract.blockHash,
                 blockNumber = contract.blockNumber,
-                name = Utils.hexToString(name),
-                symbol = Utils.hexToString(symbol),
+                name = hexToString(name),
+                symbol = hexToString(symbol),
                 decimals = decimals,
                 totalSupply = totalSupply
               )
