+++ conseil-common/src/main/scala/tech/cryptonomic/conseil/common/mantis/rpc/MantisRpcCommands.scala
@@ -2,7 +2,7 @@ package tech.cryptonomic.conseil.common.mantis.rpc
 
 import io.circe.{Decoder, Encoder, HCursor, Json}
 
-import tech.cryptonomic.conseil.common.mantis.domain.Bytecode
+import tech.cryptonomic.conseil.common.EvmDomain.Bytecode
 import tech.cryptonomic.conseil.common.rpc.RpcClient.RpcRequest
 
 /**
