+++ conseil-common/src/test/scala/tech/cryptonomic/conseil/common/EvmDomain/BytecodeTest.scala
@@ -1,4 +1,4 @@
-package tech.cryptonomic.conseil.common.ethereum.domain
+package tech.cryptonomic.conseil.common.EvmDomain
 
 import org.scalatest.wordspec.AnyWordSpec
 import org.scalatest.matchers.should.Matchers
