+++ conseil-common/src/test/scala/tech/cryptonomic/conseil/common/ethereum/EthereumFixtures.scala
@@ -4,7 +4,7 @@ import java.time.Instant
 import java.sql.Timestamp
 
 import tech.cryptonomic.conseil.common.ethereum.rpc.json._
-import tech.cryptonomic.conseil.common.ethereum.domain._
+import tech.cryptonomic.conseil.common.EvmDomain._
 
 /**
   * Fixtures for Ethereum based on MainNet block at height 0x18E70 with hash 0x017685281a11f6514538b113d62c7efb9852922ff308f4596d2c37c6f4717214
